Add storage and TimeStringUtility classes to separate methods and extract common methods and constants when dealing with LocalDateTime and LocalDate
