+++ src/main/java/dodo/Storage.java
@@ -1,0 +1,104 @@
+package dodo;
+
+import java.io.*;
+import java.time.LocalDateTime;
+import java.time.format.DateTimeParseException;
+import java.util.Scanner;
+
+import static dodo.TimeStringUtility.DTF;
+
+public class Storage {
+    private File storage;
+
+    public Storage(File storage) {
+        this.storage = storage;
+    }
+
+    public File getFile() {
+        return storage;
+    }
+
+    public void existenceCheck() {
+        try {
+            if (!storage.exists()) {
+                if (storage.getParentFile() != null) {
+                    storage.getParentFile().mkdirs();
+                }
+                storage.createNewFile();;
+            }
+        } catch(IOException ex) {
+            System.out.println(ex.getMessage());
+        }
+    }
+
+    private boolean stringToBoolean(String line) throws DodoException {
+        switch (line) {
+        case "T":
+            return true;
+        case "F":
+            return false;
+        default:
+            throw new DodoException("Incorrect done marking formatting");
+        }
+    }
+
+    public LocalDateTime stringToLdt(String line) throws DodoException {
+        LocalDateTime ldt;
+        try {
+            ldt = LocalDateTime.parse(line, DTF);
+        } catch (DateTimeParseException ex) {
+            throw new DodoException("Incorrect formatting of time. Use: yyyy-mm-dd hh:ss");
+        }
+        return LocalDateTime.parse(line, DTF);
+    }
+
+    public void readTo(TaskList tasks) throws FileNotFoundException, DodoException {
+        Scanner storageScanner = new Scanner(storage);
+        while (storageScanner.hasNextLine()) {
+            String line = storageScanner.nextLine();
+            String[] lineArr = line.split("\\|");
+            int len = lineArr.length;
+            if (len < 3 || len > 5) {
+                throw new DodoException("Incorrect storage formatting");
+            }
+            switch (lineArr[0]) {
+            case "T":
+                if (len != 3) {
+                    throw new DodoException("Incorrect storage formatting");
+                }
+                tasks.addTask(new Todo(lineArr[2], stringToBoolean(lineArr[1])));
+                break;
+            case "D":
+                if (len != 4) {
+                    throw new DodoException("Incorrect storage formatting");
+                }
+                tasks.addTask(new Deadline(lineArr[2], stringToLdt(lineArr[3]), stringToBoolean(lineArr[1])));
+                break;
+            case "E":
+                if (len != 5) {
+                    throw new DodoException("Incorrect storage formatting");
+                }
+                tasks.addTask(new Event(lineArr[2], stringToLdt(lineArr[3]), stringToLdt(lineArr[4]),
+                        stringToBoolean(lineArr[1])));
+                break;
+            default:
+                throw new DodoException("Incorrect storage formatting");
+            }
+        }
+    }
+
+    public void update(TaskList tasks) throws IOException {
+        File temp = new File(System.getProperty("user.dir") + "/data/temp.txt");
+        BufferedWriter sW = new BufferedWriter(new FileWriter(temp));
+        for (int i = 0; i < tasks.size(); i++) {
+            sW.write(tasks.get(i).getStorageString() + System.lineSeparator());
+        }
+        sW.close();
+        if (!storage.delete()) {
+            System.out.println("Cannot delete file");
+        }
+        if (!temp.renameTo(storage)) {
+            System.out.println("Cannot rename file");
+        }
+    }
+}
